Crate moonbeam_evm_tracer

Substrate EVM tracing.

The purpose of this crate is enable tracing the EVM opcode execution and will be used by both Dapp developers - to get a granular view on their transactions - and indexers to access the EVM callstack (internal transactions).

Proxies EVM messages to the host functions.

Modules

tracer